Add files via upload
Potential API to input data- MoveBank
------------------------------------------
This is an API that has records of animal migration data. There are many studies available. I have chosen one which focuses on migration of branta leucopsis, or the common barnacle goose which migrates to Ireland every Winter from Greenland, and is a high risk bird for Avian Flu.

The code has been sampled from the Open WeatherMap API example completed in class, and modified as necessary with assistance from ChatGPT to ensure smooth integration. The code will be tested during the laboratory session on Monday the 24th of November.